Add access callback and empty check in annotation_load.
The access callback was not being run when it was in annotation.store.inc.
I've added it to annotation.module to make sure that it is run.
Also added an empty check to annotation_load to remove the errors that
show up at annotation/:id when there are no annotations.